+++ apps/mobile/src/app/(home,search,profile,notifications,settings)/settings/preferences.tsx
@@ -187,6 +187,11 @@ export default function Screen() {
             key: 'blurNsfw',
             label: 'media.blurNsfw',
           },
+          {
+            icon: 'Image',
+            key: 'saveToAlbum',
+            label: 'media.saveToAlbum',
+          },
 
           null,
           t('system.title'),